+++ built/base/TreeUtils.js
@@ -206,15 +206,18 @@ class TreeUtils {
         });
         return result;
     }
+    // todo: rename
     static sortByAccessor(accessor) {
         return (objectA, objectB) => {
+            const nodeAFirst = -1;
+            const nodeBFirst = 1;
             const av = accessor(objectA);
             const bv = accessor(objectB);
-            let result = av < bv ? -1 : av > bv ? 1 : 0;
+            let result = av < bv ? nodeAFirst : av > bv ? nodeBFirst : 0;
             if (av === undefined && bv !== undefined)
-                result = -1;
+                result = nodeAFirst;
             else if (bv === undefined && av !== undefined)
-                result = 1;
+                result = nodeBFirst;
             return result;
         };
     }